requirements.txt
streamlit
insightface==0.7.3
onnxruntime
opencv-python
numpy
Pillow
